Task 02_03(AB),02_04(A),02_05(AB)  Burov

+++ src/by/it/burov/jd02_02/utils/Sleeper.java
@@ -7,9 +7,9 @@ public class Sleeper {
     private Sleeper() {
     }
 
-    public static void sleep(int millis) {
+    public static void sleep(double millis) {
         try {
-            Thread.sleep(millis / Constants.K_SPEED);
+            Thread.sleep((long) (millis / Constants.K_SPEED));
         } catch (InterruptedException e) {
             throw new StoreException(e);
         }
